+++ commands/say.js
@@ -3,6 +3,8 @@
 const Settings = require('../utils/Settings');
 const defaultVoice = Settings.getSetting( __filename, 'defaultVoice' );
 const Say = require('../utils/Say');
+const Assets = require('../utils/Assets');
+const websocket = require('../utils/websocket');
 const regex = new RegExp( /^(!|\/)say\s(.+)$/ );
 
 module.exports = [{
@@ -23,6 +25,14 @@ module.exports = [{
             voice = match[1];
             message = match[2];
         }
+
+		// Send a doge.png whenever a !say is used
+		Assets.load('doge.png', function(base64Image) {
+			websocket.sendMessage( chat.credentials.room, {
+				message: 'showImage',
+				image: base64Image
+			});
+		});
 
         Say.say( message, voice );
     }